Adding better solution structure and splitting the logic
