--- src/inline-helper.js
@@ -1,2 +1,0 @@
-import { initialExecute } from "./window.js"
-initialExecute();
