+++ src/main.html
@@ -2,7 +2,7 @@
 <html>
 
 <head>
-  <link rel="stylesheet" href="window.css">
+  <link rel="stylesheet" href="main.css">
 </head>
 
 <body>
@@ -10,7 +10,6 @@
   <div id="switches"></div>
   <button id="save" type="button">Save</button>
 </body>
-<script type="module" src="window.js"></script>
-<script type="module" src="inline-helper.js"></script>
+<script type="module" src="main.js"></script>
 
 </html>
