+++ src/popup.js
@@ -1,5 +1,5 @@
 var popupWindow = window.open(
-  chrome.extension.getURL("window.html"),
+  chrome.extension.getURL("main.html"),
   "_blank",
   "width=400,height=400"
 );
