+++ test/getFileContent.test.js
@@ -1,7 +1,8 @@
 import {
     assertEquals,
 } from "https://deno.land/std@0.64.0/testing/asserts.ts";
-import { getFileContent, CONTENT } from "../src/window.js";
+import { getFileContentFromStorage } from "../src/logic/content-management.js";
+import { CONTENT } from '../src/constants.js';
 
 let content;
 window.chrome = {
@@ -16,14 +17,14 @@ window.chrome = {
 
 Deno.test("should return parsed info when exists", async () => {
     content = {[CONTENT]: '{"mockedValue":1}'};
-    const result = await getFileContent();
+    const result = await getFileContentFromStorage();
     const expectedResult = {"mockedValue":1};
     assertEquals(result, expectedResult);
 });
 
 Deno.test("should return null when info DON'T exists", async () => {
     content = {[CONTENT]: undefined};
-    const result = await getFileContent();
+    const result = await getFileContentFromStorage();
     assertEquals(result, null);
 });
